Improve performance of React Hooks
