+++ package.json
@@ -17,7 +17,7 @@
     "@material-ui/lab": "^4.0.0-alpha.40",
     "@material-ui/styles": "^4.9.0",
     "@yusuke-suzuki/qoodish-api-js-client": "^0.2.1",
-    "@yusuke-suzuki/rize-router": "^0.1.1",
+    "@yusuke-suzuki/rize-router": "^0.1.2",
     "blueimp-canvas-to-blob": "^3.17.0",
     "blueimp-load-image": "^2.26.0",
     "express": "^4.17.1",
