+++ src/components/organisms/NavBar.js
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useMemo } from 'react';
 import { useMappedState } from 'redux-react-hook';
 
 import NavDrawer from '../molecules/NavDrawer';
@@ -9,9 +9,26 @@ import AppBar from '@material-ui/core/AppBar';
 
 import { match } from 'path-to-regexp';
 
-const NavBar = () => {
-  const [isMapDetail, setIsMapDetail] = useState(false);
+const isModalLocation = location => {
+  return location && location.state && location.state.modal;
+};
 
+const routesShowBackButton = [
+  {
+    path: '/maps/:mapId'
+  },
+  {
+    path: '/maps/:mapId/reports/:reviewId'
+  },
+  {
+    path: '/spots/:placeId'
+  },
+  {
+    path: '/users/:userId'
+  }
+];
+
+const NavBar = () => {
   const mapState = useCallback(
     state => ({
       currentLocation: state.shared.currentLocation
@@ -21,22 +38,32 @@ const NavBar = () => {
 
   const { currentLocation } = useMappedState(mapState);
 
-  useEffect(() => {
-    if (currentLocation.state && currentLocation.state.modal) {
-      return;
+  const isMapDetail = useMemo(() => {
+    if (isModalLocation(currentLocation)) {
+      return false;
     }
 
-    if (match('/maps/:mapId')(currentLocation.pathname)) {
-      setIsMapDetail(true);
-    } else {
-      setIsMapDetail(false);
-    }
+    return match('/maps/:mapId')(currentLocation.pathname);
   }, [currentLocation]);
+
+  const showBackButton = useMemo(() => {
+    if (isModalLocation(currentLocation)) {
+      return false;
+    }
+
+    return routesShowBackButton.find(route => {
+      return match(route.path)(currentLocation.pathname);
+    });
+  }, [currentLocation, routesShowBackButton]);
 
   return (
     <div>
       <AppBar position="fixed">
-        {isMapDetail ? <MapToolbar /> : <NavToolbar />}
+        {isMapDetail ? (
+          <MapToolbar />
+        ) : (
+          <NavToolbar showBackButton={showBackButton} />
+        )}
       </AppBar>
       <NavDrawer />
     </div>
